A less cubic diablo, god bless normals

+++ header/rasterizer.h
@@ -7,6 +7,6 @@
 #include "Vecteur.h"
 
 void triangle(Model &mod, Frame &frame, std::vector<Vec3f> &v, float, \
-	      float *zbuffer, std::vector<Vec3i>, std::vector<Vec3i>);
+	      float *zbuffer, std::vector<Vec3i>, std::vector<Vec3f>);
 
 #endif /* !RASTERIZER_H_ */
